requirements.txt
dj-database-url==0.5.0
Django>=3.1.2,<4
gunicorn==20.0.4
psycopg2==2.8.6
whitenoise==5.2.0
django-extensions
pydot
django-crispy-forms
djangorestframework
#django-constance[database]
#django-dynamic-preferences